로또 2단계-메서드 파라미터타입 에러 해결

+++ src/main/java/lotto/domain/Lotto.java
@@ -23,7 +23,7 @@ public class Lotto {
 		return numbers;
 	}
 	
-	public long countMatchNumbers(Numbers winNumbers) {
+	public int countMatchNumbers(Numbers winNumbers) {
 		return numbers.countMatch(winNumbers);
 	}
 }
